final debug

+++ src/components/post-form/PostForm.jsx
@@ -6,11 +6,12 @@ import { useNavigate } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 
 function PostForm({post}) {
+
     const {register, handleSubmit, watch, setValue, control, getValues} = useForm({
         defaultValues: {
-            title: post?.title || '',
-            slug: post?.slug || '',
-            content: post?.content || '',
+            title: post?.title || '' ,
+            slug: post?.slug || '' ,
+            content: post?.content || '' ,
             status: post?.status || 'active'
         }
     })
@@ -19,23 +20,21 @@ function PostForm({post}) {
     const userData = useSelector(state => state.userData)
 
     const submit = async (data) => {
+
         if (post) {
             const file = data.featuredImage[0] ? await appwriteService.uploadFile(data.featuredImage[0]) : null
-
-            if (file) {
-                await appwriteService.deleteFile(post.featuredImage)
-            }
 
             const dbPost = await appwriteService.updatePost(post.$id, {
                 ...data,
                 featuredImage: file ? file.$id : undefined
             })
+            // console.log("dbPost", dbPost);
+            if(file && dbPost) await appwriteService.deleteFile(post.featuredImage)
 
             if (dbPost) {
                 navigate(`/post/${dbPost.$id}`)
             }
         } else {
-            console.log(userData);
             const file = data.featuredImage[0] ? await appwriteService.uploadFile(data.featuredImage[0]) : null
 
             if (file) {
@@ -103,7 +102,6 @@ function PostForm({post}) {
                     label='Featured Image: '
                     type='file'
                     className='mb-4'
-                    {...register('title', {required: true})}
                     accept="image/png, image/jpg, image/jpeg, image/gif"
                     {...register('featuredImage', {required: !post})}
                 />
